feat: add functions to get fragment size, tag size and number of reads
This is quite clunky unfortunately as there is no good way of obtaining
the results of the predictd function. I have to write the stderr results
to a file and then manually find them (very human readable file, not
computer readable). This section will need updating if MACS ever changes
how this file looks.