feat: use chi squared distribution to calculate lambda ci
This is better as I am no longer counting on the number of reads to give
me a confidence interval. This was a bad approach as in places where
there was no peak, I had to add an arificial read. This caused massive
confidence intervals where lambda was small, resulting in oversized CIs.