feat: use new method for calculating lambda ci
The previous method (although better) still creates CIs that are very
large (which is given as you'd expect this anyways). This then lead to
pvalue CIs that were basically guaranteed to overlap. Instead I am
taking a different approach. The reason why we are doing this is to
account for slight fluctuations in the pileup read counts (where
slightly moved reads could result in much higher pileups). To account
for this, you can get the variance of lambdas over the region and use
this to get a standard error to use to get a CI.

This method is quite a bit better at getting what I actually want (and
is more likely to give a tighter interval). There is room to make a
sliding window approach instead getting the variance from the whole set
of lambdas.